Relocate Encoder CUDA graph manager
